Moved cleanup to own folder.
