+++ src/tree/builder/cleanup/getElementContent.js
